update and delete methods changes

+++ src/services/shoppingCart.service.js
@@ -14,12 +14,19 @@ class ShoppingCartService {
 
 	static addProduct = async (shoppingCartId, productId, SKU, quantity) => {
 
-		const currentProduct = await ShoppingCart_Product.findOne({where: {shoppingCartId, productId, SKU}});
+
+		const currentProductPromise =  ShoppingCart_Product.findOne({where: {shoppingCartId, productId, SKU}});
+		const cartPromise = ShoppingCart.findByPk(shoppingCartId);
+
+		const [currentProduct, cart] = await Promise.all([currentProductPromise, cartPromise]);
+
+		if(!cart) throw boom.notFound('Shopping cart not found');
 
 		if(currentProduct === null) {
 			const shoppingProduct = await ShoppingCart_Product.create({shoppingCartId, productId, quantity, SKU});
 			if(!(shoppingProduct instanceof ShoppingCart_Product)) throw boom.badImplementation('Unexpected error');
-		} else {
+		}
+		else {
 			const res = await ShoppingCart_Product.update({
 				quantity: currentProduct.quantity + 1,
 			}, {
@@ -32,18 +39,28 @@ class ShoppingCartService {
 	}
 
 	static updateProduct = async (shoppingCartId, ref, newData) => {
+
+		const cart = await ShoppingCart.findByPk(shoppingCartId);
+		if(!cart) throw boom.notFound('Shopping cart not found');
+
 		const res = await ShoppingCart_Product.update(newData, {
 			where: {ref}
 		});
 		if(res === null) throw boom.badImplementation('Unexpected error');
-		if(Array.isArray(res) && res[0] === 0) throw boom.badRequest("The ref or data is not valid");
+		if(Array.isArray(res) && res[0] === 0) throw boom.notFound('Item not found');
+
 		return await this.findOne(shoppingCartId);
 	}
 
 	static deleteProduct = async (shoppingCartId, ref) => {
+		const cart = await ShoppingCart.findByPk(shoppingCartId);
+		if(!cart) throw boom.notFound('Shopping cart not found');
+
 		const res = await ShoppingCart_Product.destroy({where: {ref}});
+
 		if(res === null) throw boom.badImplementation('Unexpected error');
-		if(res === 0) throw boom.badRequest("The ref is not valid");
+		if(res === 0) throw boom.notFound('Item not found');
+
 		return await this.findOne(shoppingCartId);
 	}
